Fixes issue 21: More Priority Levels
The priority levels are stored in a enum now (makes it easy to add/change priority levels).
When saving a priority a text is used instead of a number (and only if it is different compared to the default level).
When loading the new (and old nummeric) format is recognised.
The priority icons are renamed (to match the priority level names).

TODO: Create icons for HIGHEST and LOWEST level (now they are just copies of HIGH and LOW).